+++ commands/outputs.go
@@ -4,16 +4,12 @@ import (
 	"encoding/json"
 	"fmt"
 
+	"github.com/lmtani/cromwell-cli/pkg/cromwell"
 	"github.com/urfave/cli/v2"
 )
 
-type OutputsResponse struct {
-	ID      string
-	Outputs map[string]interface{}
-}
-
 func OutputsWorkflow(c *cli.Context) error {
-	cromwellClient := FromInterface(c.Context.Value("cromwell"))
+	cromwellClient := cromwell.FromInterface(c.Context.Value("cromwell"))
 	resp, err := cromwellClient.Outputs(c.String("operation"))
 	if err != nil {
 		return err
