+++ commands/submit.go
@@ -3,29 +3,18 @@ package commands
 import (
 	"fmt"
 
+	"github.com/lmtani/cromwell-cli/pkg/cromwell"
 	"github.com/urfave/cli/v2"
 	"go.uber.org/zap"
 )
 
-type SubmitResponse struct {
-	ID     string
-	Status string
-}
-
-type SubmitRequest struct {
-	workflowSource       string
-	workflowInputs       string
-	workflowDependencies string
-	workflowOptions      string
-}
-
 func SubmitWorkflow(c *cli.Context) error {
-	cromwellClient := FromInterface(c.Context.Value("cromwell"))
-	r := SubmitRequest{
-		workflowSource:       c.String("wdl"),
-		workflowInputs:       c.String("inputs"),
-		workflowDependencies: c.String("dependencies"),
-		workflowOptions:      c.String("options")}
+	cromwellClient := cromwell.FromInterface(c.Context.Value("cromwell"))
+	r := cromwell.SubmitRequest{
+		WorkflowSource:       c.String("wdl"),
+		WorkflowInputs:       c.String("inputs"),
+		WorkflowDependencies: c.String("dependencies"),
+		WorkflowOptions:      c.String("options")}
 	resp, err := cromwellClient.Submit(r)
 	if err != nil {
 		return err
